Add tabs for graphs in flight modal
